requirements.txt
flask
torch
transformers
pillow
scikit-learn
open_clip_torch
pandas
numpy
tqdm
